[all resource export] fixed export handle
# Conflicts:
#	Cargo.lock
